Computing descriptives
This version of R summary, you can describe a single variable and you can get a large range of statistics such as the N, the mean, the standard deviation, the median, the trend mean, the mdeian absolute. So a great set of statistics  that can give you  a much richer idea of what is going on with your data. I use the factor "entrepreneur" and here I can see the number of searches relatively speaking on a state by state basis for the entreoreneur term. The box represents the middle 50% of values. The dark line in the middle there is the median. It has half of the scores below, half above. the dotted lines are the whiskers,and they go out what are called the fences. both are low and high end. We can get values that are used for locating each of the elements in this graph by using the function called boxplot.stats.